# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports": {
      "get": {
        "operationId": "ReportsController_list",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportListDto"
                }
              }
            }
          },
          "400": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "리포트 목록",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/unwritten": {
      "get": {
        "operationId": "ReportsController_unwritten",
        "parameters": [
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnwrittenDto"
                }
              }
            }
          },
          "400": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "§47 안 쓴 리포트 — 강사별 밀린 건수와 예상 차감",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/reports/{serId}/{onDate}": {
      "get": {
        "operationId": "ReportsController_detail",
        "parameters": [
          {
            "name": "serId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "onDate",
            "required": true,
            "in": "path",
            "description": "REP 복합 유니크 키의 날짜",
            "schema": {
              "example": "2026-08-27",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "리포트 상세 — 입력 순서·제한도 서버 계약으로 내려준다",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{serId}/{onDate}/draft": {
      "put": {
        "operationId": "ReportsController_saveDraft",
        "parameters": [
          {
            "name": "serId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "onDate",
            "required": true,
            "in": "path",
            "description": "REP 복합 유니크 키의 날짜",
            "schema": {
              "example": "2026-08-27",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportUpsertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "리포트 임시저장 — 빈 칸을 허용한다",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{serId}/{onDate}/submit": {
      "post": {
        "operationId": "ReportsController_submit",
        "parameters": [
          {
            "name": "serId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "onDate",
            "required": true,
            "in": "path",
            "description": "REP 복합 유니크 키의 날짜",
            "schema": {
              "example": "2026-08-27",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportUpsertDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "리포트 제출 — 3개 입력을 모두 채워야 하며 정산 기준 시각을 최초 1회만 저장한다",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{serId}/{onDate}/review": {
      "post": {
        "operationId": "ReportsController_review",
        "parameters": [
          {
            "name": "serId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "onDate",
            "required": true,
            "in": "path",
            "description": "REP 복합 유니크 키의 날짜",
            "schema": {
              "example": "2026-08-27",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "공용 오류 형식. 해당 endpoint의 입력·권한·자원·DB 검증에 따라 반환될 수 있다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "제출된 리포트 승인/반려 — 반려 사유 필수, 승인 여부는 정산과 독립",
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "TACO ERP API",
    "description": "티엔아카데미 학원 운영 백오피스 — 개발 명세서 v2 기준",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "공용 오류 코드. HTTP status와 함께 판정한다."
          },
          "message": {
            "type": "string",
            "description": "공용 필터에서 정규화한 사용자 안내 문구."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "ReportStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "grade": {
            "type": "string",
            "nullable": true
          },
          "deliver": {
            "type": "boolean",
            "description": "REP_STU.deliver — 이 학생에게 전문을 전달할지"
          }
        },
        "required": [
          "id",
          "name",
          "deliver"
        ]
      },
      "ReportRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "serId": {
            "type": "number"
          },
          "date": {
            "type": "string",
            "example": "2026-08-27",
            "description": "화면에 보이는 실제 회차 날짜"
          },
          "onDate": {
            "type": "string",
            "example": "2026-08-27",
            "description": "REP · SER_OCC 식별자인 원래 날짜"
          },
          "startMin": {
            "type": "number"
          },
          "subKey": {
            "type": "string",
            "nullable": true
          },
          "kindKey": {
            "type": "string"
          },
          "teacherId": {
            "type": "number",
            "nullable": true
          },
          "teacherName": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "enum": [
              "na",
              "plan",
              "none",
              "draft",
              "wait",
              "ok",
              "rej"
            ]
          },
          "written": {
            "type": "boolean",
            "description": "썼는가 — 정산 조건 (D-R7)"
          },
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportStudentDto"
            }
          },
          "minutesSinceEnd": {
            "type": "number",
            "description": "수업이 끝난 뒤 지난 분. 음수면 아직 안 끝났다"
          },
          "penalty": {
            "type": "number",
            "description": "지금 확정하면 깎이는 금액 (D-R32)"
          }
        },
        "required": [
          "id",
          "serId",
          "date",
          "onDate",
          "startMin",
          "kindKey",
          "state",
          "written",
          "students",
          "minutesSinceEnd",
          "penalty"
        ]
      },
      "ReportListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportRowDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UnwrittenByTeacherDto": {
        "type": "object",
        "properties": {
          "teacherId": {
            "type": "number"
          },
          "teacherName": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "oldestDate": {
            "type": "string",
            "nullable": true,
            "description": "가장 오래된 것"
          },
          "over1h": {
            "type": "number",
            "description": "1시간 넘긴 건수"
          },
          "over4h": {
            "type": "number",
            "description": "4시간 넘긴 건수"
          },
          "penalty": {
            "type": "number",
            "description": "예상 차감 합계"
          }
        },
        "required": [
          "teacherId",
          "teacherName",
          "count",
          "over1h",
          "over4h",
          "penalty"
        ]
      },
      "UnwrittenDto": {
        "type": "object",
        "properties": {
          "byTeacher": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnwrittenByTeacherDto"
            }
          },
          "total": {
            "type": "number"
          },
          "penaltyTotal": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportRowDto"
            }
          }
        },
        "required": [
          "byTeacher",
          "total",
          "penaltyTotal",
          "items"
        ]
      },
      "ReportBodyDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 2000
          },
          "progress": {
            "type": "string",
            "maxLength": 2000
          },
          "homework": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "content",
          "progress",
          "homework"
        ]
      },
      "ReportFieldDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "content",
              "progress",
              "homework"
            ]
          },
          "label": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          },
          "min": {
            "type": "number"
          },
          "max": {
            "type": "number"
          }
        },
        "required": [
          "key",
          "label",
          "hint",
          "min",
          "max"
        ]
      },
      "ReportExportFileDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "number"
          },
          "fileName": {
            "type": "string",
            "example": "20260827_김민준_고2_수학_16:30.png"
          },
          "plainText": {
            "type": "string",
            "description": "클립보드와 RSEND.body가 공유하는 서버 생성 5섹션 본문"
          }
        },
        "required": [
          "studentId",
          "fileName",
          "plainText"
        ]
      },
      "ReportDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "serId": {
            "type": "number"
          },
          "date": {
            "type": "string",
            "example": "2026-08-27",
            "description": "화면에 보이는 실제 회차 날짜"
          },
          "onDate": {
            "type": "string",
            "example": "2026-08-27",
            "description": "REP · SER_OCC 식별자인 원래 날짜"
          },
          "startMin": {
            "type": "number"
          },
          "subKey": {
            "type": "string",
            "nullable": true
          },
          "kindKey": {
            "type": "string"
          },
          "teacherId": {
            "type": "number",
            "nullable": true
          },
          "teacherName": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "enum": [
              "na",
              "plan",
              "none",
              "draft",
              "wait",
              "ok",
              "rej"
            ]
          },
          "written": {
            "type": "boolean",
            "description": "썼는가 — 정산 조건 (D-R7)"
          },
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportStudentDto"
            }
          },
          "minutesSinceEnd": {
            "type": "number",
            "description": "수업이 끝난 뒤 지난 분. 음수면 아직 안 끝났다"
          },
          "penalty": {
            "type": "number",
            "description": "지금 확정하면 깎이는 금액 (D-R32)"
          },
          "body": {
            "$ref": "#/components/schemas/ReportBodyDto"
          },
          "fields": {
            "description": "화면이 순서·문구·제한을 재정의하지 않고 그대로 그린다",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportFieldDto"
            }
          },
          "canEdit": {
            "type": "boolean",
            "description": "현재 사용자·회차·상태 기준 저장 가능 여부"
          },
          "canReview": {
            "type": "boolean",
            "description": "현재 사용자·상태 기준 승인/반려 가능 여부"
          },
          "canExport": {
            "type": "boolean",
            "description": "저장된 전문을 현재 사용자가 PNG·본문으로 출력할 수 있는지"
          },
          "exportFiles": {
            "description": "서버가 정한 학생별 PNG 파일명. canExport=false면 빈 배열",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportExportFileDto"
            }
          },
          "canDeliver": {
            "type": "boolean",
            "description": "현재 사용자가 학부모 전달 이력을 만들 수 있는지. manager 이상만 true"
          },
          "subjectName": {
            "type": "string",
            "description": "전문·파일명에 함께 쓰는 서버 과목명"
          },
          "lang": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "nullable": true
          },
          "submittedAt": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "nullable": true
          },
          "rejectReason": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "serId",
          "date",
          "onDate",
          "startMin",
          "kindKey",
          "state",
          "written",
          "students",
          "minutesSinceEnd",
          "penalty",
          "body",
          "fields",
          "canEdit",
          "canReview",
          "canExport",
          "exportFiles",
          "canDeliver",
          "subjectName",
          "lang"
        ]
      },
...
      "ReportUpsertDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 2000
          },
          "progress": {
            "type": "string",
            "maxLength": 2000
          },
          "homework": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "content",
          "progress",
          "homework"
        ]
      },
      "ReportReviewDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ]
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "description": "반려 시 필수 (D-R13)"
          }
        },
        "required": [
          "decision"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}